+++ src/screens/SignIn/components/GoogleSignIn/index.tsx
@@ -1,0 +1,10 @@
+import { GoogleContainer } from "./styles";
+
+
+export default function GoogleSignIn(){
+    return(
+        <GoogleContainer>
+            
+        </GoogleContainer>
+    )
+}