+++ src/screens/SignIn/components/GoogleSignIn/styles.ts
@@ -1,0 +1,5 @@
+import styled from "styled-components/native";
+
+export const GoogleContainer = styled.View`
+    width: 100%;
+`